Successful rerouting.
