+++ dinosaurs.json
@@ -1,1 +1,1 @@
-[{"name":"Karen","type":"Unicornisaur"},{"name":"Petrie","type":"Pterosaur"},{"name":"Littlefoot","type":"Apatosaur"},{"name":"Test","type":"Testtype"}]
+[{"name":"Petrie","type":"Pterosaur"},{"name":"Littlefoot","type":"Apatosaur"},{"name":"Test","type":"Testtype"},{"name":"Terry","type":"Pterosaur"},{"name":"Meow","type":"Cat"}]
