+++ prehistoric_creatures.json
@@ -1,1 +1,1 @@
-[{"type":"mastodon","img_url":"https://cdn-images-1.medium.com/max/1200/1*a2VvYsKGApR-E1SnT5O7yQ.jpeg"},{"type":"saber-toothed salmon","img_url":"https://cottagelife.com/wp-content/uploads/2014/11/Oncorhynchus_rastrosus.jpg"},{"type":"megalonyx","img_url":"https://animalgeography.files.wordpress.com/2018/08/sloth-banner-e1535192925361.jpg"},{"type":"Nick Cage","img_url":"https://png.pngitem.com/pimgs/s/125-1256517_nicholas-cage-face-png-picture-library-download-nicolas.png"}]
+[{"type":"saber-toothed salmon","img_url":"https://cottagelife.com/wp-content/uploads/2014/11/Oncorhynchus_rastrosus.jpg"},{"type":"megalonyx","img_url":"https://animalgeography.files.wordpress.com/2018/08/sloth-banner-e1535192925361.jpg"},{"type":"Nicolas Cage","img_url":"https://png.pngitem.com/pimgs/s/125-1256517_nicholas-cage-face-png-picture-library-download-nicolas.png"},{"type":"Mack the Knife","img_url":"https://helpx.adobe.com/content/dam/help/en/photoshop/using/convert-color-image-black-white/jcr_content/main-pars/before_and_after/image-before/Landscape-Color.jpg"}]
