+++ views/dinosaurs/edit.ejs
@@ -1,4 +1,4 @@
-<form method="POST" action="/dinosaurs/<%= dinoId %>/?_method=PUT">
+<form method="POST" action="/dinosaurs/dinosaurs/<%= dinoId %>/?_method=PUT">
     <input class="text-field" type="text" value="<%= dino.name %>" name="name"/>
     <input class="text-field" type="text" value="<%= dino.type %>" name="type"/>
     <div id="submit-box">
