+++ views/dinosaurs/index.ejs
@@ -4,10 +4,10 @@
             <%= dino.name %> is a <%= dino.type %>.
                 <!-- <a id="edit-button" href="/dinosaurs/edit/<%= index %>">EDIT</a> -->
                 <div id="button-box">
-                    <form action="/dinosaurs/edit/<%= index %>">
+                    <form action="/dinosaurs/dinosaurs/edit/<%= index %>">
                         <input class="submit-button" type="submit" value="EDIT"/>
                     </form>
-                    <form method="POST" action="/dinosaurs/<%= index %>/?_method=DELETE">
+                    <form method="POST" action="/dinosaurs/dinosaurs/<%= index %>/?_method=DELETE">
                             <input class="submit-button" type="submit" value="DELETE"/>
                     </form>
                 </div>
