+++ views/dinosaurs/new.ejs
@@ -1,4 +1,4 @@
-<form method="post" action="/dinosaurs">
+<form method="post" action="/dinosaurs/dinosaurs">
     <input class="text-field" id="dinoName" type="text" name="name" placeholder="name"/>
     <input class="text-field" id="dinoType" type="text" name="type" placeholder="type"/>
     <div id="submit-box">
