+++ views/layout.ejs
@@ -14,16 +14,16 @@
                 <a href="/">Home</a>
             </li>
             <li id="links">
-                <a href="/dinosaurs">Dinosaurium</a>
+                <a href="/dinosaurs/dinosaurs">Dinosaurium</a>
             </li>
             <li id="links">
-                <a href="/dinosaurs/new">New Dinosaur</a>
+                <a href="/dinosaurs/dinosaurs/new">New Dinosaur</a>
             </li>
             <li id="links">
-                <a href="/prehistoric_creatures">Prehistorium</a>
+                <a href="/prehistoric_creatures/prehistoric_creatures">Prehistorium</a>
             </li>
             <li id="links">
-                <a href="/prehistoric_creatures/new">New Prehistoric Creature</a>
+                <a href="/prehistoric_creatures/prehistoric_creatures/new">New Prehistoric Creature</a>
             </li>
         </ul>
     </nav>
