+++ views/prehistoric_creatures/edit.ejs
@@ -1,4 +1,4 @@
-<form method="POST" action="/prehistoric_creatures/<%= creatureID %>/?_method=PUT">
+<form method="POST" action="/prehistoric_creatures/prehistoric_creatures/<%= creatureID %>/?_method=PUT">
     <input class="text-field" type="text" value="<%= creature.type %>" name="type"/>
     <input class="text-field" type="text" value="<%= creature.img_url %>" name="img_url"/>
     <div id="submit-box">
