+++ views/prehistoric_creatures/index.ejs
@@ -6,10 +6,10 @@
                 <img src= <%= creature.img_url %>>
             </div>
                 <div id="button-box">
-                    <form action="/prehistoric_creatures/edit/<%= index %>">
+                    <form action="/prehistoric_creatures/prehistoric_creatures/edit/<%= index %>">
                         <input class="submit-button" type="submit" value="EDIT"/>
                     </form>
-                    <form method="POST" action="/prehistoric_creatures/<%= index %>/?_method=DELETE">
+                    <form method="POST" action="/prehistoric_creatures/prehistoric_creatures/<%= index %>/?_method=DELETE">
                             <input class="submit-button" type="submit" value="DELETE"/>
                     </form>
                 </div>
