+++ views/prehistoric_creatures/new.ejs
@@ -1,4 +1,4 @@
-<form method="post" action="/prehistoric_creatures">
+<form method="post" action="/prehistoric_creatures/prehistoric_creatures">
     <input class="text-field" id="creatureType" type="text" name="type" placeholder="type"/>
     <input class="text-field" id="creatureImg" type="text" name="img_url" placeholder="http://img_url"/>
     <div id="submit-box">
